requirements.txt
discord-py-interactions
python-dotenv
requests
pyairtable
